+++ pxl2D/src/graphics/PXL_ImageUtils.cpp
@@ -58,7 +58,7 @@ PXL_Bitmap* PXL_load_png(std::string file_name, PXL_Bitmap* bitmap) {
 
 	const PXL_uint row_length = bitmap->get_width() * ((bit_depth * channels) / 8);
 	png_bytep* row_pointers = new png_bytep[bitmap->get_height()];
-	for (int y = 0; y < bitmap->get_height(); ++y) {
+	for (size_t y = 0; y < bitmap->get_height(); ++y) {
 		row_pointers[y] = (png_bytep)(bitmap->get_pixels() + (y * row_length));
 	}
